pyproject.toml
[project]
name = "telegram-scraper"
version = "0.1.0"
description = "Telegram chat scraper with optional PostgreSQL persistence"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Pavlos Labada" }
]
license = { text = "MIT" }

dependencies = [
  "telethon>=1.33,<2.0",
  "psycopg2-binary>=2.9,<3.0",
]

[tool.black]
line-length = 88
target-version = ["py310"]
skip-string-normalization = true

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"